改bug mp

+++ src/main/java/com/demo/community/controller/CommentController.java
@@ -52,7 +52,7 @@ public class CommentController {
 
     @ResponseBody
     @RequestMapping(value = "/comment/{id}",method = RequestMethod.GET)
-    public ResultDTO<List<CommentDTO>> comments(@PathVariable(name = "id") int id){
+    public ResultDTO<List<CommentDTO>> comments(@PathVariable(name = "id") Long id){
         List<CommentDTO> commentDTOS = commentService.listByParentId(id, CommentTypeEnum.COMMENT,0);
         return ResultDTO.okOf(commentDTOS);
     }
